+++ client/src/main/java/org/evosuite/testsmells/smells/AssertionRoulette.java
@@ -42,10 +42,10 @@ public class AssertionRoulette extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
-        int count = 0;
-        int numStatements = 0;
+        long count = 0;
+        long numStatements = 0;
 
         Statement currentStatement;
