+++ client/src/main/java/org/evosuite/testsmells/smells/DuplicateAssert.java
@@ -42,9 +42,9 @@ public class DuplicateAssert extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
-        int count = 0;
+        long count = 0;
 
         Set<Triple<Method, Class<? extends Assertion>, Object>> tripleSet = new LinkedHashSet<>();
