+++ client/src/main/java/org/evosuite/testsmells/smells/UnusedInputs.java
@@ -39,9 +39,9 @@ public class UnusedInputs extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
-        int count = 0;
+        long count = 0;
 
         Statement currentStatement;
